# docs/gochat-openapi3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GoChat API",
    "version": "1.0.0",
    "description": "根据项目当前 Gin 路由、controller、DTO 和 service 结构整理的 OpenAPI 3 文档。所有 REST 接口虽然基本都返回 HTTP 200，但真实业务结果请以响应体中的 code 字段为准：200=成功，400=业务错误，500=系统错误。WebSocket 接口为 GET /wss?client_id={userId}。"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "README 中默认本地开发后端地址"
    }
  ],
...
  "paths": {
...
    "/user/wsLogout": {
      "post": {
        "tags": [
          "WebSocket"
        ],
        "summary": "WebSocket 登出",
        "operationId": "wsLogout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WsLogoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登出结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/wss": {
      "get": {
        "tags": [
          "WebSocket"
        ],
        "summary": "建立 WebSocket 连接",
        "operationId": "wsLogin",
        "description": "建立连接时必须带 client_id 查询参数，例如 ws://localhost:8000/wss?client_id=U202607010001。连接建立后前端向服务端发送 ChatMessageRequest JSON。服务端会先推送一条纯文本欢迎消息，随后按消息类型推送 MessageItem 或 AVMessageItem JSON。",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "description": "当前登录用户 id",
            "schema": {
              "type": "string",
              "example": "U202607010001"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols，升级为 WebSocket"
          },
          "200": {
            "description": "当 client_id 缺失时，接口会直接返回普通 JSON 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataResponse"
                }
              }
            }
          }
        },
        "x-apipost-websocket": {
          "sendMessageSchema": {
            "$ref": "#/components/schemas/ChatMessageRequest"
          },
          "receiveMessageSchemas": [
            {
              "$ref": "#/components/schemas/MessageItem"
            },
            {
              "$ref": "#/components/schemas/AVMessageItem"
            }
          ]
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BaseResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码：200=成功，400=业务错误，500=系统错误",
            "enum": [
              200,
              400,
              500
            ],
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          }
        }
      },
      "NoDataResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          }
        ]
      },
...
      "WsLogoutRequest": {
        "type": "object",
        "required": [
          "owner_id"
        ],
        "properties": {
          "owner_id": {
            "type": "string",
            "example": "U202607010001"
          }
        }
      },
...
      "ChatMessageRequest": {
        "type": "object",
        "required": [
          "session_id",
          "type",
          "send_id",
          "send_name",
          "send_avatar",
          "receive_id"
        ],
        "properties": {
          "session_id": {
            "type": "string",
            "example": "S202607010001"
          },
          "type": {
            "type": "integer",
            "description": "消息类型：0=文本，1=语音（预留），2=文件，3=音视频信令",
            "enum": [
              0,
              1,
              2,
              3
            ],
            "example": 0
          },
          "content": {
            "type": "string",
            "description": "文本消息时使用",
            "example": "你好"
          },
          "url": {
            "type": "string",
            "description": "文件消息时使用，通常为 /static/files/xxx",
            "example": "/static/files/demo.pdf"
          },
          "send_id": {
            "type": "string",
            "example": "U202607010001"
          },
          "send_name": {
            "type": "string",
            "example": "alice"
          },
          "send_avatar": {
            "type": "string",
            "example": "/static/avatars/avatar_u1001.png"
          },
          "receive_id": {
            "type": "string",
            "description": "接收用户或群聊 id",
            "example": "U202607010002"
          },
          "file_size": {
            "type": "string",
            "example": "120KB"
          },
          "file_type": {
            "type": "string",
            "example": "pdf"
          },
          "file_name": {
            "type": "string",
            "example": "demo.pdf"
          },
          "av_data": {
            "type": "string",
            "description": "当 type=3 时传入 JSON 字符串，结构可参考 AVDataPayload",
            "example": "{\"messageId\":\"PROXY\",\"type\":\"start_call\"}"
          }
        }
      },
...
      "MessageItem": {
        "type": "object",
        "properties": {
          "send_id": {
            "type": "string",
            "example": "U202607010001"
          },
          "send_name": {
            "type": "string",
            "example": "alice"
          },
          "send_avatar": {
            "type": "string",
            "example": "/static/avatars/avatar_u1001.png"
          },
          "receive_id": {
            "type": "string",
            "example": "U202607010002"
          },
          "type": {
            "type": "integer",
            "description": "0=文本，1=语音，2=文件，3=音视频信令",
            "enum": [
              0,
              1,
              2,
              3
            ],
            "example": 0
          },
          "content": {
            "type": "string",
            "example": "你好"
          },
          "url": {
            "type": "string",
            "example": "/static/files/demo.pdf"
          },
          "file_type": {
            "type": "string",
            "example": "pdf"
          },
          "file_name": {
            "type": "string",
            "example": "demo.pdf"
          },
          "file_size": {
            "type": "string",
            "example": "120KB"
          },
          "created_at": {
            "type": "string",
            "example": "2026-07-01 12:00:00"
          }
        }
      },
      "AVMessageItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MessageItem"
          },
          {
            "type": "object",
            "properties": {
              "av_data": {
                "type": "string",
                "example": "{\"messageId\":\"PROXY\",\"type\":\"start_call\"}"
              }
            }
          }
        ]
      },
...
    }
  }
}